% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhgis_ts_data.R
\name{define_nhgis_ts_extract}
\alias{define_nhgis_ts_extract}
\title{Define a NHGIS time series extract}
\usage{
define_nhgis_ts_extract(
  year = NULL,
  tables = NULL,
  geography = c("county", "state"),
  extent = "us",
  output = c("tidy", "wide", "file"),
  shape_year = NULL,
  basis = 2008,
  geometry = FALSE,
  ...,
  time_series_tables = NULL,
  description = NULL,
  shapefiles = NULL,
  data_format = "csv_no_header",
  validate = TRUE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\arguments{
\item{output}{Used to set \code{tst_layout} value. c("tidy", "wide", "file")}

\item{extract}{If not \code{NULL}, extract is returned as is.}
}
\description{
Define a NHGIS time series extract
}
\keyword{internal}
