% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ipumsr_extract_paths.R
\name{download_ipumsr_extract}
\alias{download_ipumsr_extract}
\title{Download IPUMS extract using \code{ipumsr::wait_for_extract} and
\code{ipumsr::download_extract}}
\usage{
download_ipumsr_extract(
  extract = NULL,
  download_dir = getwd(),
  overwrite = FALSE,
  progress = TRUE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\arguments{
\item{extract}{One of:
\itemize{
\item An \code{\link[ipumsr:ipums_extract-class]{ipums_extract}} object
\item The data collection and extract number formatted as a string of the
form \code{"collection:number"} or as a vector of the form
\code{c("collection", number)}
\item An extract number to be associated with your default IPUMS
collection. See \code{\link[ipumsr:set_ipums_default_collection]{set_ipums_default_collection()}}
}

For a list of codes used to refer to each collection, see
\code{\link[ipumsr:ipums_data_collections]{ipums_data_collections()}}.}

\item{download_dir}{Path to the directory where the files should be written.
Defaults to current working directory.}

\item{overwrite}{If \code{TRUE}, overwrite any conflicting files that
already exist in \code{download_dir}. Defaults to \code{FALSE}.}

\item{progress}{If \code{TRUE}, output progress bar showing the status of the
download request. Defaults to \code{TRUE}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of the \code{IPUMS_API_KEY} environment variable. See
\code{\link[ipumsr:set_ipums_api_key]{set_ipums_api_key()}}.}
}
\description{
Download IPUMS extract using \code{ipumsr::wait_for_extract} and
\code{ipumsr::download_extract}
}
