% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gisjoin.R
\name{filter_nhgisjoin}
\alias{filter_nhgisjoin}
\title{Filter a NHGIS data frame by a state FIPS and county FIPS numbers}
\usage{
filter_nhgisjoin(
  data,
  state = NULL,
  county = NULL,
  values = NULL,
  gisjoin_col = "NHGISJOIN"
)
}
\arguments{
\item{values}{Optional vector of GISJOIN values. state_fips and county_fips
are ignored if this is supplied.}
}
\description{
By default, filter to counties in the Baltimore metro area.
}
\keyword{internal}
