% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt_nhgis_places}
\alias{fmt_nhgis_places}
\title{Format NHGIS places data by joining additional attributes and adding a label
column}
\usage{
fmt_nhgis_places(
  data,
  label_col = "label",
  name_col = "NAME",
  remove_pattern = " city$"
)
}
\arguments{
\item{label_col}{Label column name.}

\item{name_col}{Name column to use as the basis for the label.}

\item{remove_pattern}{Passed to \code{pattern} for \code{\link[stringr:str_remove]{stringr::str_remove()}}}
}
\description{
\code{\link[=fmt_nhgis_places]{fmt_nhgis_places()}} joins Census division, region, and USPS abbreviation
columns based on the state value and creates a new label column following the
pattern of "\if{html}{\out{<NAME column with remove pattern applied>}}, \if{html}{\out{<USPS State
Abbreviation>}}".
}
\keyword{fmt}
\keyword{internal}
