% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_ipumsr.R
\name{fmt_perc_value_col}
\alias{fmt_perc_value_col}
\title{Format a percent value column with rounding and handling for 0 and NA values}
\usage{
fmt_perc_value_col(
  data,
  value_col = "value",
  denominator_prefix = "denominator_",
  denominator_data = NULL,
  ...,
  perc_prefix = "perc_",
  digits = 2
)
}
\arguments{
\item{data}{Data to use in creating the percent value column.}

\item{denominator_data}{Optional data frame with prepared denominator data.}

\item{...}{Additional parameters passed to \code{dplyr::left_join}.}
}
\description{
\code{\link[=fmt_perc_value_col]{fmt_perc_value_col()}} is an internal utility function for create a percent
value column based on a specified value column and denominator column
(following the name convention of the prefix + value column).
}
\keyword{internal}
