% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_ipumsr_data.R
\name{get_ipumsr_cache}
\alias{get_ipumsr_cache}
\title{Get cached object from \code{{ipumseasyr}} package cache}
\usage{
get_ipumsr_cache(code, refresh = FALSE, file = NULL, path = ipumsr_cache_dir())
}
\arguments{
\item{refresh}{If \code{FALSE}, do not read a file from cache. If \code{TRUE}, read a
file from cache if it exists at the supplied path.}

\item{file}{File name of the cached RDS file.}

\item{path}{Path to package cache directory. Defaults to \code{\link[=ipumsr_cache_dir]{ipumsr_cache_dir()}}}
}
\description{
\code{\link[=get_ipumsr_cache]{get_ipumsr_cache()}} reads a cached IPUMS data file using \code{\link[readr:read_rds]{readr::read_rds()}}.
}
\keyword{internal}
