% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ipumsr_extract_paths.R
\name{get_ipumsr_extract_paths}
\alias{get_ipumsr_extract_paths}
\title{Get extract paths for extract with optional support for cached extract files}
\usage{
get_ipumsr_extract_paths(
  extract = NULL,
  data_file = NULL,
  shape_file = NULL,
  submit_extract = TRUE,
  download_extract = TRUE,
  download_dir = getwd(),
  overwrite = FALSE,
  progress = TRUE,
  refresh = FALSE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\arguments{
\item{extract}{An \code{\link[ipumsr:ipums_extract-class]{ipums_extract}} object.}

\item{submit_extract}{If \code{extract} is not \code{NULL} and \code{submit_extract = TRUE},
use \link[ipumsr:submit_extract]{ipumsr::submit_extract} to submit the extract.}

\item{api_key}{API key associated with your user account. Defaults to the
value of the \code{IPUMS_API_KEY} environment variable. See
\code{\link[ipumsr:set_ipums_api_key]{set_ipums_api_key()}}.}
}
\value{
A named list with "data" and "shape" elements containing extract file paths.
}
\description{
Download extract with \code{\link[=download_ipumsr_extract]{download_ipumsr_extract()}} then return a list of file
paths.
}
