% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extract_history.R
\name{get_nhgis_extract_history}
\alias{get_nhgis_extract_history}
\title{Get NHGIS Extract history}
\usage{
get_nhgis_extract_history(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ipumsr:get_extract_history]{ipumsr::get_extract_history}}
  \describe{
    \item{\code{collection}}{Character string of the IPUMS collection for which to
retrieve extract history. Defaults to the current default
collection, if it exists. See \code{\link[ipumsr:set_ipums_default_collection]{set_ipums_default_collection()}}.

For a list of codes used to refer to each collection, see
\code{\link[ipumsr:ipums_data_collections]{ipums_data_collections()}}.}
    \item{\code{how_many}}{The number of extract requests for which to retrieve
information. Defaults to the 10 most recent extracts.}
    \item{\code{delay}}{Number of seconds to delay between
successive API requests, if multiple requests are needed to retrieve all
records.

A delay is highly unlikely to be necessary and is
intended only as a fallback in the event that you cannot retrieve your
extract history without exceeding the API rate limit.}
    \item{\code{api_key}}{API key associated with your user account. Defaults to the
value of the \code{IPUMS_API_KEY} environment variable. See
\code{\link[ipumsr:set_ipums_api_key]{set_ipums_api_key()}}.}
  }}
}
\description{
Get NHGIS Extract history
}
\keyword{internal}
