% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipumseasyr-package.R
\docType{package}
\name{ipumseasyr-package}
\alias{ipumseasyr}
\alias{ipumseasyr-package}
\title{ipumseasyr: Easy Access to IPUMS Data}
\description{
A package with helper functions extending the {ipumsr} package for accessing NHGIS and other IMPUMS data sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/elipousson/ipumseasyr}
  \item \url{https://elipousson.github.io/ipumseasyr/}
  \item Report bugs at \url{https://github.com/elipousson/ipumseasyr/issues}
}

}
\author{
\strong{Maintainer}: Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [copyright holder]

}
\keyword{internal}
