% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs_nhgis.R
\name{labs_nhgis}
\alias{labs_nhgis}
\title{Helper function for labelling ggplot2 plots with the appropriate credit for NHGIS data}
\usage{
labs_nhgis(
  ...,
  caption = NULL,
  credit = "IPUMS NHGIS, University of Minnesota, www.nhgis.org.",
  prefix = "Source: ",
  width = 80
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggplot2:labs]{ggplot2::labs}}
  \describe{
    \item{\code{title}}{The text for the title.}
    \item{\code{subtitle}}{The text for the subtitle for the plot which will be
displayed below the title.}
    \item{\code{caption}}{The text for the caption which will be displayed in the
bottom-right of the plot by default.}
    \item{\code{tag}}{The text for the tag label which will be displayed at the
top-left of the plot by default.}
    \item{\code{alt,alt_insight}}{Text used for the generation of alt-text for the plot.
See \link[ggplot2]{get_alt_text} for examples.}
  }}
}
\description{
Helper function for labelling ggplot2 plots with the appropriate credit for NHGIS data
}
\keyword{internal}
