% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_nhgis_shapefiles.R
\name{list_nhgis_shapefiles}
\alias{list_nhgis_shapefiles}
\title{List NHGIS Shapefiles}
\usage{
list_nhgis_shapefiles(
  geography = c("county", "state"),
  year,
  extent = "us",
  basis = 2008,
  validate = FALSE,
  ...,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\description{
List NHGIS Shapefiles
}
\keyword{internal}
