% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhgis_ts_data.R
\name{list_nhgis_ts_tables}
\alias{list_nhgis_ts_tables}
\title{List NHGIS time series tables using \code{ipumsr::get_metadata_nhgis}}
\usage{
list_nhgis_ts_tables(
  ...,
  cache = TRUE,
  cache_file = "nhgis_time_series_tables.rds",
  refresh = FALSE,
  integration = NULL,
  recursive = TRUE
)
}
\arguments{
\item{...}{Additional parameters passed to \code{ipumsr::get_metadata_nhgis()}}

\item{refresh}{If \code{FALSE}, do not read a file from cache. If \code{TRUE}, read a
file from cache if it exists at the supplied path.}

\item{integration}{Optional filter for geographical integration.}
}
\description{
Use \code{ipumsr::get_metadata_nhgis()} with \code{type = "time_series_tables"} to
return a data frame of time series tables. Optionally filter by geographical
integration type "nominal" or "standardized" ("2010" or "standardized to
2010" also work).
}
