% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_nhgis_shapefiles.R
\name{lookup_nhgis_state}
\alias{lookup_nhgis_state}
\title{Lookup NHGIS State Code}
\usage{
lookup_nhgis_state(state)
}
\arguments{
\item{state}{State name, abbreviation, FIPS code, or NHGIS code}
}
\description{
Lookup NHGIS State Code
}
\keyword{internal}
