% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhgis_ts_tables}
\alias{nhgis_ts_tables}
\title{NHGIS Time Series Table names}
\format{
A character vector with 389 time series table names.
}
\usage{
nhgis_ts_tables
}
\description{
A vector of NHGIS time series table names named with table descriptions.
}
\keyword{datasets}
