% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nhgis_files.R
\name{read_ipums_geometry}
\alias{read_ipums_geometry}
\title{Read IPUMS geometry using \code{ipumsr::read_ipums_sf}}
\usage{
read_ipums_geometry(
  shape_file = NULL,
  path = NULL,
  file_select = NULL,
  vars = "GISJOIN",
  encoding = NULL,
  bind_multiple = TRUE,
  add_layer_var = NULL,
  verbose = FALSE
)
}
\description{
@inheritParams ipumsr::read_ipums_sf
@export
@importFrom ipumsr read_ipums_sf
}
