% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nhgis_files.R
\name{read_nhgis_files}
\alias{read_nhgis_files}
\title{Read NHGIS data and geometry}
\usage{
read_nhgis_files(
  path = NULL,
  data_file = NULL,
  data_file_select = NULL,
  shape_file = NULL,
  shape_file_select = NULL,
  verbose = FALSE,
  geometry = FALSE,
  ...
)
}
\description{
@inheritParams ipumsr::read_nhgis
@export
@importFrom dplyr left_join join_by
}
