% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_ipumsr.R
\name{summarise_denominator}
\alias{summarise_denominator}
\title{Summarise data using a summary function}
\usage{
summarise_denominator(
  data,
  .by = c("GISJOIN", "YEAR"),
  .f = function(x) {
     sum(x, na.rm = TRUE)
 },
  denominator_column_title_value = "Total",
  variable_col = "variable",
  value_col = "value",
  column_title_col = "column_title",
  denominator_prefix = "denominator_"
)
}
\arguments{
\item{data}{Input data w/ column names including \code{value_col} and any columns specified in \code{.by}.}

\item{.by}{Join columns used by \code{dplyr::summarise}}

\item{.f}{Function to apply to column specified by \code{value_col}.}
}
\description{
In some cases, a denominator value can be created by summarizing across a set
of variables for a GISJOIN and year value. If a denominator variable is not
already identified it may need to be calculated and joined with
\code{\link[=fmt_perc_value_col]{fmt_perc_value_col()}} setting \code{drop_existing = TRUE} before using this function
}
\keyword{internal}
