% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usa_states}
\alias{usa_states}
\title{U.S. States Reference data}
\format{
A data frame with 50 rows and 4 variables:
\describe{
\item{\code{STATE}}{State name}
\item{\code{STUSPS}}{State USPS abbreviation}
\item{\code{division}}{U.S. Census Division name}
\item{\code{region}}{U.S. Census Region name}
}
}
\usage{
usa_states
}
\description{
Reference data with U.S. state names, USPS abbreviations, and Census
divisions, and regions.
}
\keyword{datasets}
