% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhgis_ts_data.R
\name{define_nhgis_ts_extract}
\alias{define_nhgis_ts_extract}
\title{Define a NHGIS time series extract using \code{ipumsr::define_extract_nhgis}}
\usage{
define_nhgis_ts_extract(
  year = NULL,
  tables = NULL,
  geography = c("county", "state"),
  extent = "us",
  output = c("tidy", "wide", "file"),
  shape_year = NULL,
  basis = 2008,
  geometry = FALSE,
  ...,
  time_series_tables = NULL,
  description = NULL,
  shapefiles = NULL,
  data_format = "csv_no_header",
  validate = TRUE,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\arguments{
\item{output}{Used to set \code{tst_layout} value. c("tidy", "wide", "file")
corresponding to "time_by_row_layout", "time_by_column_layout", or
"time_by_file_layout".}

\item{geometry}{If \code{TRUE}, include shapefiles in the defined extract. If
shapefiles is \code{NULL}, the function uses \code{\link[=list_nhgis_shapefiles]{list_nhgis_shapefiles()}} with
\code{shape_year} as the \code{year} parameter.}

\item{...}{
  Arguments passed on to \code{\link[ipumsr:define_extract_nhgis]{ipumsr::define_extract_nhgis}}
  \describe{
    \item{\code{datasets}}{List of dataset specifications for any
\href{https://www.nhgis.org/overview-nhgis-datasets}{datasets}
to include in the extract request. Use \code{\link[ipumsr:ds_spec]{ds_spec()}} to create a
\code{ds_spec} object containing a dataset specification. See examples.}
    \item{\code{geographic_extents}}{Vector of geographic extents to use for
all of the \code{datasets} in the extract definition (for instance, to obtain
data within a particular state). Use \code{"*"} to select all available extents.

Required when any of the \code{datasets} included in the extract definition
include \code{geog_levels} that require extent selection. See
\code{\link[ipumsr:get_metadata_nhgis]{get_metadata_nhgis()}} to determine if a geographic level requires extent
selection. At the time of writing, NHGIS supports extent selection only
for blocks and block groups.}
    \item{\code{breakdown_and_data_type_layout}}{The desired layout
of any \code{datasets} that have multiple data types or breakdown values.
\itemize{
\item \code{"single_file"} (default) keeps all data types and breakdown values in
one file
\item \code{"separate_files"} splits each data type or breakdown value into its
own file
}

Required if any \code{datasets} included in the extract definition consist of
multiple data types (for instance, estimates and margins of error) or have
multiple breakdown values specified. See \code{\link[ipumsr:get_metadata_nhgis]{get_metadata_nhgis()}} to
determine whether a requested dataset has multiple data types.}
  }}

\item{time_series_tables}{List of time series table specifications for any
\href{https://www.nhgis.org/time-series-tables}{time series tables}
to include in the extract request. Use \code{\link[ipumsr:tst_spec]{tst_spec()}} to create a
\code{tst_spec} object containing a time series table specification. See
examples.}

\item{description}{Description of the extract.}

\item{shapefiles}{Names of any \href{https://www.nhgis.org/gis-files}{shapefiles}
to include in the extract request.}

\item{data_format}{The desired format of the extract data file.
\itemize{
\item \code{"csv_no_header"} (default) includes only a minimal header in the first
row
\item \code{"csv_header"} includes a second, more descriptive header row.
\item \code{"fixed_width"} provides data in a fixed width format
}

Note that by default, \code{\link[ipumsr:read_nhgis]{read_nhgis()}} removes the additional header row in
\code{"csv_header"} files.

Required when an extract definition includes any \code{datasets} or
\code{time_series_tables}.}

\item{api_key}{API key associated with your user account. Defaults to the
value of the \code{IPUMS_API_KEY} environment variable. See
\code{\link[ipumsr:set_ipums_api_key]{set_ipums_api_key()}}.}
}
\description{
\code{\link[=define_nhgis_ts_extract]{define_nhgis_ts_extract()}} is a wrapper for \code{\link[ipumsr:define_extract_nhgis]{ipumsr::define_extract_nhgis()}}
with defaults that support the creation of tidy data using
\code{\link[=read_nhgis_data]{read_nhgis_data()}} or \code{\link[=pivot_nhgis_data]{pivot_nhgis_data()}}.
}
