% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhgis_ts_data.R
\name{get_nhgis_ts_data}
\alias{get_nhgis_ts_data}
\title{Get NHGIS time series data}
\usage{
get_nhgis_ts_data(
  year = NULL,
  tables = NULL,
  geography = c("county", "state"),
  extent = "us",
  output = c("tidy", "wide", "file"),
  basis = 2008,
  shape_year = NULL,
  geometry = FALSE,
  extract = NULL,
  data_file = NULL,
  shape_file = NULL,
  state = NULL,
  ...,
  time_series_tables = NULL,
  description = NULL,
  shapefiles = NULL,
  data_format = "csv_no_header",
  validate = TRUE,
  submit_extract = TRUE,
  download_extract = TRUE,
  read_files = TRUE,
  download_dir = getwd(),
  overwrite = FALSE,
  progress = TRUE,
  verbose = progress,
  api_key = Sys.getenv("IPUMS_API_KEY")
)
}
\arguments{
\item{output}{Used to set \code{tst_layout} value. c("tidy", "wide", "file")
corresponding to "time_by_row_layout", "time_by_column_layout", or
"time_by_file_layout".}

\item{geometry}{If \code{TRUE}, include shapefiles in the defined extract. If
shapefiles is \code{NULL}, the function uses \code{\link[=list_nhgis_shapefiles]{list_nhgis_shapefiles()}} with
\code{shape_year} as the \code{year} parameter.}

\item{extract}{An \code{\link[ipumsr:ipums_extract-class]{ipums_extract}} object.}

\item{data_file}{Path to a .zip archive containing an NHGIS extract or
a single file from an NHGIS extract.}

\item{shape_file}{Path to a single .shp file or a .zip archive
containing at least one .shp file. See Details section.}

\item{...}{Additional parameters passed to both \code{\link[=read_nhgis_data]{read_nhgis_data()}} and
\code{\link[=read_ipums_geometry]{read_ipums_geometry()}}.}

\item{time_series_tables}{List of time series table specifications for any
\href{https://www.nhgis.org/time-series-tables}{time series tables}
to include in the extract request. Use \code{\link[ipumsr:tst_spec]{tst_spec()}} to create a
\code{tst_spec} object containing a time series table specification. See
examples.}

\item{description}{Description of the extract.}

\item{shapefiles}{Names of any \href{https://www.nhgis.org/gis-files}{shapefiles}
to include in the extract request.}

\item{data_format}{The desired format of the extract data file.
\itemize{
\item \code{"csv_no_header"} (default) includes only a minimal header in the first
row
\item \code{"csv_header"} includes a second, more descriptive header row.
\item \code{"fixed_width"} provides data in a fixed width format
}

Note that by default, \code{\link[ipumsr:read_nhgis]{read_nhgis()}} removes the additional header row in
\code{"csv_header"} files.

Required when an extract definition includes any \code{datasets} or
\code{time_series_tables}.}

\item{download_dir}{Path to the directory where the files should be written.
Defaults to current working directory.}

\item{overwrite}{If \code{TRUE}, overwrite any conflicting files that
already exist in \code{download_dir}. Defaults to \code{FALSE}.}

\item{progress}{If \code{TRUE}, output progress bar showing the status of the
download request. Defaults to \code{TRUE}.}

\item{verbose}{Logical controlling whether to display output when loading
data. If \code{TRUE}, displays IPUMS conditions, a progress bar, and
column types. Otherwise, all are suppressed.

Will be overridden by \code{readr.show_progress} and \code{readr.show_col_types}
options, if they are set.}

\item{api_key}{API key associated with your user account. Defaults to the
value of the \code{IPUMS_API_KEY} environment variable. See
\code{\link[ipumsr:set_ipums_api_key]{set_ipums_api_key()}}.}
}
\description{
Use \code{define_nhgis_ts_extract()}, \code{ipumsr::submit_extract()},
\code{ipumsr::download_extract()}, and \code{read_nhgis_files()} to define, submit,
download, and read a NHGIS time series extract. This function is \emph{only}
recommended for interactive use and is \emph{not} recommended if you are
requesting a large number of tables or geographies.
}
