% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_ipumsr_data.R
\name{ipumsr_cache_dir}
\alias{ipumsr_cache_dir}
\title{Get path to \code{{ipumseasyr}} package cache directory}
\usage{
ipumsr_cache_dir(pkg = "ipumseasyr")
}
\description{
Get path to \code{{ipumseasyr}} package cache directory
}
