% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_ipumsr.R
\name{join_nhgis_percent_change}
\alias{join_nhgis_percent_change}
\title{Join a percent change in variable relative to a reference year}
\usage{
join_nhgis_percent_change(
  data,
  reference_year = NULL,
  value_col = "value",
  reference_prefix = "reference_",
  gisjoin_col = "NHGISCODE",
  variable_col = "variable",
  year_col = "YEAR",
  rank_col = "rank",
  rank = NULL,
  rank_n = NULL,
  ...,
  perc_prefix = "perc_change_",
  digits = 2
)
}
\arguments{
\item{reference_year}{Reference year to use when calculating a percent change
column.}

\item{rank, rank_n}{Passed to \code{x} and \code{n} arguments of \code{\link[dplyr:ntile]{dplyr::ntile()}} to
join a reference rank value.}
}
\description{
\code{\link[=join_nhgis_percent_change]{join_nhgis_percent_change()}} joins a percent change column relative to a
reference year. Optionally join a rank from the reference year using
\code{\link[dplyr:ntile]{dplyr::ntile()}}.
}
