% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs_nhgis.R
\name{labs_nhgis}
\alias{labs_nhgis}
\title{Label ggplot2 plots with the appropriate credit caption for NHGIS data}
\usage{
labs_nhgis(
  ...,
  caption = NULL,
  credit = "IPUMS NHGIS, University of Minnesota, www.nhgis.org.",
  prefix = "Source: ",
  collapse = " ",
  width = 80
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggplot2:labs]{ggplot2::labs}}
  \describe{
    \item{\code{title}}{The text for the title.}
    \item{\code{subtitle}}{The text for the subtitle for the plot which will be
displayed below the title.}
    \item{\code{tag}}{The text for the tag label which will be displayed at the
top-left of the plot by default.}
    \item{\code{dictionary}}{A named character vector to serve as dictionary.
Automatically derived labels, such as those based on variables will
be matched with \code{names(dictionary)} and replaced by the matching
entry in \code{dictionary}.}
    \item{\code{alt,alt_insight}}{Text used for the generation of alt-text for the plot.
See \link[ggplot2]{get_alt_text} for examples. \code{alt} can also be a function that
takes the plot as input and returns text as output. \code{alt} also accepts
rlang \link[rlang:as_function]{lambda} function notation.}
  }}

\item{caption}{Optional text appearing before IPUMS credit text.}

\item{credit}{Credit line for IPUMS.}

\item{prefix}{Prefix text appearing before \code{credit} text.}

\item{collapse}{String to collapse \code{caption} and \code{credit}. Defaults to \code{" "}.
Set to \code{"\\n"} to place the credit line on a separate line following the
caption. Ignored if caption is \code{NULL}.}

\item{width}{Maximum width of caption line passed to \code{\link[stringr:str_wrap]{stringr::str_wrap()}}.}
}
\description{
\code{\link[=labs_nhgis]{labs_nhgis()}} adds a standard credit caption for NHGIS data to make consistent attribution easier.
}
\keyword{ggplot2}
