% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_ipumsr.R
\name{pivot_nhgis_data}
\alias{pivot_nhgis_data}
\alias{join_nhgis_percent}
\title{Pivot NHGIS data longer to assign denominator variables and join percent values}
\usage{
pivot_nhgis_data(
  data,
  variable_col = "variable",
  value_col = "value",
  column_title_col = "column_title",
  denominator_prefix = "denominator_",
  cols_vary = "slowest"
)

join_nhgis_percent(
  data,
  variable_col = "variable",
  value_col = "value",
  column_title_col = "column_title",
  denominator_prefix = "denominator_",
  perc_prefix = "perc_",
  join_cols = c("GISJOIN", "YEAR"),
  digits = 2
)
}
\arguments{
\item{data}{A data frame to pivot.}

\item{variable_col}{Variable column name}

\item{value_col}{Value column name}

\item{column_title_col}{Column title column name (to be created from column labels)}

\item{cols_vary}{When pivoting \code{cols} into longer format, how should the
output rows be arranged relative to their original row number?
\itemize{
\item \code{"fastest"}, the default, keeps individual rows from \code{cols} close
together in the output. This often produces intuitively ordered output
when you have at least one key column from \code{data} that is not involved in
the pivoting process.
\item \code{"slowest"} keeps individual columns from \code{cols} close together in the
output. This often produces intuitively ordered output when you utilize
all of the columns from \code{data} in the pivoting process.
}}

\item{perc_prefix}{Prefix string to use for calculated percent value.}

\item{label_col}{Label column name}
}
\description{
\code{\link[=pivot_nhgis_data]{pivot_nhgis_data()}} uses \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}} switches NHGIS data from a
wide to long format and creates a denominator column based on a preset
crosswalk between variables and corresponding denominators. The input data
must have variable column labels present.

\code{\link[=join_nhgis_percent]{join_nhgis_percent()}} uses the denominator values added in pivoting to
calculating a percent share value. This feature is supported for many but not
all of the most popular NHGIS time series variables. Both function uses a
similar set of conventions as the \code{{getACS}} package to support ease of code
reuse between NHGIS and American Community Survey (ACS) data.

At presnet, all numeric columns that do not appear to be an identifier are
pivoted.
}
\keyword{internal}
