% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_ipumsr.R
\name{pivot_nhgis_data}
\alias{pivot_nhgis_data}
\alias{join_nhgis_percent}
\title{Pivot NHGIS data longer to assign denominator variables}
\usage{
pivot_nhgis_data(
  data,
  variable_col = "variable",
  value_col = "value",
  label_col = "label",
  denominator_prefix = "denominator_",
  variable_starts_with = c("A", "B", "D0", "AV", "A4", "BS", "BUQ", "B0J", "B7", "CV",
    "CM", "CL")
)

join_nhgis_percent(
  data,
  variable_col = "variable",
  value_col = "value",
  label_col = "label",
  denominator_prefix = "denominator_",
  perc_prefix = "perc_"
)
}
\arguments{
\item{variable_col}{Variable column name}

\item{value_col}{Value column name}

\item{label_col}{Label column name}

\item{perc_prefix}{Prefix string to use for calculated percent value.}
}
\description{
\code{\link[=pivot_nhgis_data]{pivot_nhgis_data()}} switches NHGIS data from a wide to long format and
creates a denominator column based on a preset crosswalk between variables
and corresponding denominators.
}
\keyword{internal}
