% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nhgis_files.R
\name{read_nhgis_data}
\alias{read_nhgis_data}
\title{Read NHGIS data using \code{ipumsr::read_nhgis}}
\usage{
read_nhgis_data(
  data_file = NULL,
  path = NULL,
  file_select = NULL,
  verbose = FALSE,
  ...
)
}
\description{
@inheritParams ipumsr::read_nhgis
@export
@importFrom vctrs vec_rbind
@importFrom ipumsr read_nhgis
}
